requirements.txt
fastapi==0.85.2
uvicorn==0.19.0
optimum-graphcore>=0.6.0, <0.7.0
matplotlib
examples-utils[common] @ git+https://github.com/graphcore/examples-utils@latest_stable
ipywidgets>=7,<8